Rename module: entrydate -> datetime
